unimed/requirements.txt
pandas
python-docx
odfpy
PyPDF2
num2words